requirements.txt
ultralytics
manga-ocr
simple-lama-inpainting
deep-translator
streamlit
Pillow